Updated code to create delivery entry in the database.

+++ ApproveRequest.php
@@ -88,10 +88,55 @@ if (!$result) {
     echo json_encode($response);
     die();
 }
+$RequestRow = mysqli_fetch_assoc($result);
 
-// TODO: Figure out what we want todo once the request is confirmed
-// For now we will just delete the request
+// Move the request to the deliveries table
 
+// Generate a delivery ID
+// Generate Unique ID for this request
+// Generate a UserID For The User.
+$DeliveryID ="";
+$numrows =1;
+while ($numrows != 0) {
+    // Keep Generating a Random Number Until We Can Find One That Hasnt Been Used Yet.
+    $DeliveryID ="";
+    for ($i = 0; $i<8; $i++) {
+        $DeliveryID  .= mt_rand(0,9);
+    }
+
+    // Select Any Data From The Match Database To See If We Have Already Used The Match ID.
+    $result = mysqli_query($conn,"SELECT delivery_id FROM deliveries WHERE delivery_id = '$DeliveryID'");
+    if (!$result) {
+        $response['success'] = FALSE;
+        $response['debug'] = "Delivery ID generation failed. ".mysqli_error($conn);
+        echo json_encode($response);
+        die();
+    }
+    $numrows = mysqli_num_rows($result);
+}
+
+// Get the data we need from the request
+$senderID = $RequestRow['senderID'];
+$receiverID = $RequestRow['revceiverID'];
+$message = $RequestRow['packageMessage'];
+
+// Generate a date/time stamp
+date_default_timezone_set('UTC');
+$date = date("F j, Y, g:i:s a");
+
+// Inser the data into the deliveries database
+$MoveQuery = "INSERT INTO deliveries (delivery_id, drone_id, sender_id, receiver_id, package_message, delivery_status, time_stamp) VALUES ('$DeliveryID', '1234', '$senderID', '$receiverID', '$message', 'holding for takeoff', '$date')";
+$result = mysqli_query($conn, $MoveQuery);
+
+if (!$result) {
+    // MYSQL Error
+    $response['success'] = FALSE;
+    $response['debug'] = "MYSQL query failed: ".mysqli_error($conn);
+    echo json_encode($response);
+    die();
+}
+
+// Delete it from the request from the database
 $DeleteQuery = "DELETE FROM requests where requestID='$requestID'";
 $result = mysqli_query($conn, $DeleteQuery);
 if (!$result) {
